+++ Math/TriangularNumbers.hs
@@ -1,0 +1,16 @@
+
+-- | Triangular numbers and various helper functions.
+--
+-- Main use is for linearization of triangular array indexing.
+-- 
+-- Triangular numbers:
+-- @
+-- T_n = Σ_{k=1)^n k = 1 + 2 + 3 + ... + n =
+--
+-- n * (n+1) / 2 = (n+1) `choose` 2
+-- @
+--
+--
+
+module Math.TriangularNumbers where
+