+++ changelog.md
@@ -1,3 +1,9 @@
+0.0.0.2
+-------
+
+- Math.TriangularNumbers collects functions for triangular numbers and indexing
+- Data.Combined.Foldable
+
 0.0.0.1
 -------